requirements.txt
# 音频处理
ffmpeg-python==0.2.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# 语音识别和AI
openai==1.3.7
anthropic==0.7.8
speech-recognition==3.10.0
whisper-openai==20231117

# 实时流处理
pyaudio==0.2.11
websockets==12.0
asyncio-mqtt==0.16.1

# 字幕处理
srt==3.5.3
webvtt-py==0.4.6

# 视频处理
opencv-python==4.8.1.78
moviepy==1.0.3

# 工具库
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
loguru==0.7.2

# 配置和序列化
pydantic==2.5.0
pyyaml==6.0.1
toml==0.10.2

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
